Move logic that handles editor options into editMod
Previously, some of that logic was in `editMod()` and some (the part
that conditionally set `allowSelfIntersection = FALSE` for 2-D
features) was in `editFeatures.sf()`.

Migrating this logic involved one complication. Whereas in
`editFeatures.sf()` one can directly query the `sf` object being
edited for its dimension, in the farther downstream `editMod()`, that
object is no longer available. By then, the `sf` object has been
processed (in `editFeatures.sf()`) by `leafem::addFeatures()` which
adds the features to the map object. As a result, in `editMod()`, I
now test for the presence of calls to `addPolylines()` or
`addPolygons()` among the `methods` elements in `leafmap$x$calls`. I
believe that covers all of the possibilities that were formerly caught
by `if(any(sf::st_dimension(x) == 2))`, but I have not exhaustively
tested this.

This commit also replaces repeated use of `editor[1]` with an
equivalent but clearer initial call to `editor <- match.arg(editor)`.